refactor(luascript): move lua api to own file
