+++ src/lua/variant.h
@@ -1,7 +1,10 @@
 #ifndef FS_LUAVARIANT_H
 #define FS_LUAVARIANT_H
 
-#include "position.h"
+#include "../position.h"
+
+#include <string>
+#include <variant>
 
 enum LuaVariantType_t
 {
