+++ vc18/atlas.vcxproj
@@ -162,6 +162,7 @@
     <ClCompile Include="..\src\item.cpp" />
     <ClCompile Include="..\src\items.cpp" />
     <ClCompile Include="..\src\luascript.cpp" />
+    <ClCompile Include="..\src\lua\api.cpp" />
     <ClCompile Include="..\src\lua\env.cpp" />
     <ClCompile Include="..\src\lua\error.cpp" />
     <ClCompile Include="..\src\lua\meta.cpp" />
@@ -258,10 +259,11 @@
     <ClInclude Include="..\src\items.h" />
     <ClInclude Include="..\src\lockfree.h" />
     <ClInclude Include="..\src\luascript.h" />
-    <ClInclude Include="..\src\luavariant.h" />
+    <ClInclude Include="..\src\lua\api.h" />
     <ClInclude Include="..\src\lua\env.h" />
     <ClInclude Include="..\src\lua\error.h" />
     <ClInclude Include="..\src\lua\meta.h" />
+    <ClInclude Include="..\src\lua\variant.h" />
     <ClInclude Include="..\src\mailbox.h" />
     <ClInclude Include="..\src\map.h" />
     <ClInclude Include="..\src\matrixarea.h" />
